# specs/openapi.yaml
openapi: 3.1.0
info:
  title: ShortFlix API
  version: 0.1.1
  summary: Cross-cultural short-form video discovery — REST + A2A + MCP boundaries.
  description: |
    ShortFlix is a Progressive Web App that publishes a daily nine-card weekly of cross-cultural
    short-form video. This document covers three contract surfaces:

    1. **REST API** (consumed by the PWA at `/app` and the landing at `/`).
       Served by the **orchestrator** Cloud Run service.
    2. **A2A (Agent-to-Agent) contracts** between the four Cloud Run agent services
       (orchestrator, curator, unified-search, trend-safety).
    3. **MCP tool contracts** for the three Cloud Run MCP wrappers around RapidAPI
       (yt-shorts-mcp, ig-reels-mcp, tiktok-mcp). MCP itself is JSON-RPC 2.0; this document
       describes the HTTP-layer entry plus the JSON-RPC method shapes referenced from the
       `mcp` tag.

    ## Cross-cutting requirements (apply unless overridden per-endpoint)

    - **Auth**: OAuth 2.1 with PKCE (Authorization Code + PKCE; no implicit/password flows).
      The Google authorization step is performed at Google's endpoint; the ShortFlix server
      accepts the code at `POST /api/auth/google` and mints a session JWT delivered as an
      HttpOnly + Secure + SameSite=Lax cookie. Bearer tokens are also accepted via
      `Authorization: Bearer <jwt>` for non-cookie clients.
    - **Scopes**: `read:today`, `read:search`, `write:quest`, `write:event`, `admin:bench`.
...
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/today:
    get:
      tags: [today]
      summary: Get today's 9-pick issue for the authenticated user's locale.
      description: |
        Returns the curator's 9-pick issue for the date corresponding to "now" in the user's
        locale. Supports `If-None-Match` for ETag-based caching: when the client's ETag
        matches the current issue, 304 is returned with no body.
      operationId: getTodayIssue
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - in: header
          name: If-None-Match
          schema: { type: string }
          description: ETag from a previous response; server may return 304.
      security:
        - sessionCookie: [read:today]
        - bearerAuth:    [read:today]
        - oauth2:        [read:today]
      responses:
        '200':
          description: Today's issue.
          headers:
            ETag:
              description: SHA-256-derived strong ETag for the issue.
              schema: { type: string }
            Cache-Control:
              schema: { type: string, example: "private, max-age=120" }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '304': { description: Not Modified — caller's ETag matches. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/today/{date}:
    get:
      tags: [today]
      summary: Get a historical issue (or page across multiple days).
      description: |
        Returns the issue for `date` (UTC YYYY-MM-DD) in the authenticated user's locale.
        Cursor-paginated when fetching multiple days via the `cursor` parameter.
      operationId: getHistoricalIssue
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - in: path
          name: date
          required: true
          description: Issue date in UTC (YYYY-MM-DD).
          schema:
            type: string
            format: date
            example: "2026-05-01"
        - in: query
          name: cursor
          schema: { type: string }
          description: Opaque pagination cursor returned by a previous response.
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 30, default: 1 }
      security:
        - sessionCookie: [read:today]
        - bearerAuth:    [read:today]
        - oauth2:        [read:today]
      responses:
        '200':
          description: Historical issue page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuePage'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ───────────────────────────── SEARCH ─────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sf_session
      description: |
        HttpOnly + Secure session cookie minted after `/api/auth/google` exchange.
        OAuth 2.1 PKCE is required at the upstream (Google) leg.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer JWT containing scope claims. Used by non-cookie clients (CLI, demo recording
        tooling). Same JWT contents as the session cookie.
    oauth2:
      type: oauth2
      description: |
        OAuth 2.1 with PKCE. The authorization endpoint is Google's; the ShortFlix server
        accepts the resulting authorization code at `POST /api/auth/google`. The `tokenUrl`
        below is the **first-party** endpoint that exchanges the code for a ShortFlix
        session JWT. PKCE (S256) is mandatory — implicit and password flows are not offered.
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://api.shortflix.run.app/api/auth/google
          refreshUrl: https://api.shortflix.run.app/api/auth/refresh
          scopes:
            'read:today': Read today's issue, streak, quest state, ablation table.
            'read:search': Run unified searches.
            'write:quest': Mark quests complete; update streak.
            'write:event': Record pick interaction events.
            'admin:bench': (Internal) Trigger bench harness re-runs; not for end users.
    serviceAccount:
      type: http
      scheme: bearer
      bearerFormat: GoogleSignedIdToken
      description: |
        Service-to-service auth. Cloud Run validates the Google-signed identity token; the
        caller's service account must have `roles/run.invoker` on the target service.
  parameters:
    AcceptLanguage:
      in: header
      name: Accept-Language
      required: false
      schema: { type: string, example: "ko, en;q=0.9" }
      description: |
        BCP-47 language preference per RFC 7231 §5.3.5. Server resolves to UI chrome locale
        (en | ko at v0.1; falls back to en). Affects `Problem.title`, `Problem.detail`, and
        the `labelKey`-resolved display strings on the QuestProgressItem schema.
...
  schemas:
...
    Issue:
      type: object
      required: [id, issueDate, locale, etag, picks, noveltyScore, publishedAt]
      description: A daily issue — exactly nine picks for one (date, locale) tuple.
      properties:
        id: { type: string }
        issueDate: { type: string, format: date }
        locale: { type: string, example: "en" }
        etag: { type: string, description: SHA-256-derived ETag. }
        publishedAt: { type: string, format: date-time }
        noveltyScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Aggregate novelty score (0..1). Exposed in the ablation reveal.
        picks:
          type: array
          minItems: 9
          maxItems: 9
          description: Exactly 9 picks, ordered by `position`.
          items: { $ref: '#/components/schemas/Pick' }
    IssuePage:
      type: object
      required: [items]
      description: Cursor-paginated page of historical issues.
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Issue' }
        nextCursor:
          type: string
          nullable: true
          description: Cursor for the next page; null when no more pages.
    Pick:
      type: object
      required:
        - id
        - position
        - platform
        - videoUrl
        - thumbnailUrl
        - durationSec
        - region
        - sourceLang
        - altText
        - blurb
        - rationale
        - noveltyScore
      description: |
        One of the 9 cards in an Issue. `altText` is REQUIRED and must describe literal
        visual content (not marketing copy) for screen-reader users (a11y).
      properties:
        id: { type: string }
        position:
          type: integer
          minimum: 1
          maximum: 9
          description: 1-indexed position within the issue.
        platform:
          $ref: '#/components/schemas/Platform'
        videoUrl: { type: string, format: uri }
        thumbnailUrl: { type: string, format: uri }
        durationSec: { type: integer, minimum: 1, maximum: 600 }
        region:
          type: string
          minLength: 2
          maxLength: 2
          description: ISO 3166-1 alpha-2 of the content's apparent origin.
          example: "NG"
        sourceLang:
          type: string
          description: BCP-47 tag of the spoken/text content of the video.
          example: "yo"
        captionLang:
          type: string
          nullable: true
          description: BCP-47 tag of provided captions; null if no captions exist.
        altText:
          type: string
          minLength: 1
          maxLength: 500
          description: |
            Accessibility description for screen readers. REQUIRED. Describes literal visual
            content (e.g., "Two cooks tossing noodles on a wok over a wood fire at night")
            — NOT editorial blurb or marketing copy.
        blurb:
          type: string
          minLength: 1
          maxLength: 800
          description: Editorial 1-paragraph blurb in the source language (anti-ai voice).
        rationale:
          type: string
          minLength: 1
          maxLength: 280
          description: Curator agent's one-sentence rationale (shown in the detail modal).
        noveltyScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Per-pick novelty (0..1).
        trendSafetyTags:
          type: array
          items: { type: string }
          description: Tags emitted by the trend-safety agent (e.g., "safe", "trend_emerging").
        agentTraceId:
          type: string
          nullable: true
          description: OpenTelemetry trace id for this pick's reasoning chain.
    Platform:
      type: string
      enum: [YOUTUBE_SHORTS, INSTAGRAM_REELS, TIKTOK]
      description: Which short-form platform this content originated on.

    # ── Search ─────────────────────────────────────────────
...
    Problem:
      type: object
      description: |
        RFC 7807 problem details object. `title` and `detail` are localized per the
        request's Accept-Language; the `lang` field declares which BCP-47 locale the
        server actually rendered (which may be a fallback if the requested locale was
        unsupported).
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: 'URI reference identifying the problem class. Use "about:blank" if none.'
          example: "https://shortflix.run.app/problems/idempotency-conflict"
        title:
          type: string
          description: Short, human-readable summary. Localized.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Localized human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference to the specific occurrence.
        traceId:
          type: string
          description: OpenTelemetry trace id correlated with this error.
        lang:
          type: string
          description: BCP-47 locale that title/detail were rendered in (e.g. "en", "ko").
          example: "en"
        errors:
          type: array
          description: Optional field-level error details (validation problems).
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              code: { type: string }
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated, but missing required scope or insufficient permission.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: Too many requests.
      headers:
        RateLimit-Limit:    { schema: { type: integer } }
        RateLimit-Remaining: { schema: { type: integer } }
        RateLimit-Reset:    { schema: { type: integer } }
        Retry-After:        { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
security:
  - sessionCookie: []